pyproject.toml
[project]
name = "rio-electoral-data-2024"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"polars>=1.11.0",
	"pyyaml>=6.0.0",
	"rich>=13.7.0",
	"requests>=2.31.0",
]
